api defined

+++ src/main/java/com/bhaskar/dao/EmployeeDao.java
@@ -3,6 +3,6 @@ package com.bhaskar.dao;
 import com.bhaskar.model.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeDao extends CrudRepository<Employee, Long> {
+public interface EmployeeDao extends CrudRepository<Employee, Integer> {
 
 }
