+++ src/main/java/com/bhaskar/dao/LeaveBalanceDao.java
@@ -3,5 +3,5 @@ package com.bhaskar.dao;
 import com.bhaskar.model.LeaveBalance;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LeaveBalanceDao extends CrudRepository<LeaveBalance, Long> {
+public interface LeaveBalanceDao extends CrudRepository<LeaveBalance, Integer> {
 }
