+++ src/main/java/com/bhaskar/dao/LeaveDao.java
@@ -3,6 +3,6 @@ package com.bhaskar.dao;
 import com.bhaskar.model.Leave;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LeaveDao extends CrudRepository<Leave, Long> {
+public interface LeaveDao extends CrudRepository<Leave, Integer> {
 
 }
