+++ src/main/java/com/bhaskar/dao/LeaveStatusDao.java
@@ -3,5 +3,5 @@ package com.bhaskar.dao;
 import com.bhaskar.model.LeaveStatus;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LeaveStatusDao extends CrudRepository<LeaveStatus, Long> {
+public interface LeaveStatusDao extends CrudRepository<LeaveStatus, Integer> {
 }
